Working on simulating 100k events per strip for Prasanna.
Major changes:

1. UserDT (barely used before) changed to UserStep which determines
the simulation step size in cm.

2. WriteTraces() method now only writes the MUSIC traces in a
subdirectory the TTree is written in Simulate or
GenerateTraceDatabase.

3. New method CheckMemoryUsage() prevents Simulate() to go over 90% of
total memory. Need to implement in GenerateTraceDatabase(). This
method needs the gSystem pointer (see below).

4. New method SetROOTSystemPointer(), a.k.a ggSystem, allows
CheckMemoryUsage to work.

5. Propagare method changed from double** to int, now the DeltaE array
is not returned, instead the DelatE arrays (for the various particles)
are passed as arguments to this method.